Porting and enable emWin 5.22 function
This emWin function only add a TOUCH_Sample example.
The touch function is also enabled and configed to support emWin.
TouchPanel.c/.h are the touch driver and the touch.c/.h are the
touch interface for emWin. glcd.c/.h are the driver of LCD and the
interface between LCD and emWin is modified in LCDConf.c.

Signed-off-by: Zhang Bo <bo.zhang.zh@gmail.com>