Add the interface between CanFestival and board driver
Rename the timer.h/.c of CanFestival, because these file name are used
in other folder as other function. It will cause compile error if don't
change these file name.
This porting is an example of CanFestival. It is designed to transfer
file from PC/Master to this board(Slave).

Signed-off-by: Zhang Bo <bo.zhang.zh@gmail.com>